+++ src/presentation/theme_manager.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <QString>
+
+class QApplication;
+
+namespace rawviewer::presentation {
+
+class ThemeManager {
+public:
+    static void apply(QApplication& application, const QString& theme);
+};
+
+} // namespace rawviewer::presentation